Split up big classes, move a bit more to implementation, fix a lot of wording.
