+++ library/data/recipe/build.gradle
@@ -9,9 +9,9 @@ qslModule {
 	moduleDependencies {
 		core {
 			api("qsl_base")
-			api("resource_loader")
 			testmodOnly("qsl_base")
 			testmodOnly("lifecycle_events")
+			testmodOnly("resource_loader")
 		}
 	}
 }
